Refactored deposit by moving its methods into the deposit component rather than inside of manage
