+++ src/pages/manage/deposit/deposit.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import {BankTransferService} from "../../../providers/bank.service";
 
 @Component({
   selector: 'app-deposit',
@@ -7,9 +8,50 @@ import { Component, OnInit } from '@angular/core';
 })
 export class DepositComponent implements OnInit {
 
-  constructor() { }
+  constructor(
+    public bankSrv: BankTransferService
+  ) { }
 
   ngOnInit() {
+  }
+
+  //******************************************************
+  // AddMoney
+  //******************************************************
+
+  addMoney = {
+    currency: null,
+    amount: 0,
+    decimalPlaces: 0
+  };
+
+  public setDecimalPlaces() {
+    let minorUnit = this.addMoney.currency.minor_unit;
+    if (minorUnit == 0) {
+      this.addMoney.decimalPlaces = 1;
+    } else {
+      this.addMoney.decimalPlaces = 1.0 / Math.pow(10, minorUnit);
+    }
+  }
+
+  public submitAddMoney() {
+    this.bankSrv.deposit(this.addMoney.amount, this.addMoney.currency.code)
+      .subscribe(res => {
+        console.log(res);
+        /**
+         * Relavent information
+         * data.transactions.description
+         * data.transactions.invoice_number
+         * data.create_time
+         * data.id
+         * data.links[i].href
+         *    0 - GET
+         *    1 - REDIRECT
+         *    2 - POST
+         */
+        // let paypalUrl = res.links[1].href;
+        // let browser = new InAppBrowser(paypalUrl, '_blank', 'location=yes');
+      });
   }
 
 }
