+++ src/pages/manage/manage.ts
@@ -5,6 +5,7 @@ import { BankTransferService } from '../../providers/bank.service';
 import { Storage } from "@ionic/storage";
 import { WalletsService } from "../../providers/wallet.service";
 import {WithdrawComponent} from "./withdraw/withdraw.component";
+import {DepositComponent} from "./deposit/deposit.component";
 
 @Component({
   selector: 'page-manage',
@@ -19,6 +20,7 @@ export class ManagePage {
   loader: Loading;
 
   withdraw: WithdrawComponent;
+  deposit: DepositComponent;
 
   constructor(
     public navCtrl: NavController,
@@ -39,56 +41,19 @@ export class ManagePage {
       this.bankSrv,
       this.walletSrv
     );
+    this.deposit = new DepositComponent(
+      this.bankSrv
+    );
   }
 
   ionViewDidLoad() {
     this.storage.get('currencies')
       .then(currencies => {
         this.currencies = currencies;
-        this.addMoney.currency = this.currencies[0];
-        this.setDecimalPlaces();
+        this.deposit.addMoney.currency = this.currencies[0];
+        this.deposit.setDecimalPlaces();
       });
     this.wallets = this.walletSrv.wallets
-  }
-
-
-  //******************************************************
-  // AddMoney
-  //******************************************************
-
-  addMoney = {
-    currency: null,
-    amount: 0,
-    decimalPlaces: 0
-  };
-
-  public setDecimalPlaces() {
-    let minorUnit = this.addMoney.currency.minor_unit;
-    if (minorUnit == 0) {
-      this.addMoney.decimalPlaces = 1;
-    } else {
-      this.addMoney.decimalPlaces = 1.0 / Math.pow(10, minorUnit);
-    }
-  }
-
-  public submitAddMoney() {
-    this.bankSrv.deposit(this.addMoney.amount, this.addMoney.currency.code)
-      .subscribe(res => {
-        console.log(res);
-        /**
-         * Relavent information
-         * data.transactions.description
-         * data.transactions.invoice_number
-         * data.create_time
-         * data.id
-         * data.links[i].href
-         *    0 - GET
-         *    1 - REDIRECT
-         *    2 - POST
-         */
-        // let paypalUrl = res.links[1].href;
-        // let browser = new InAppBrowser(paypalUrl, '_blank', 'location=yes');
-      });
   }
 
 }
